Extracting pretalx fetch also in a binary to be executed in the website folder
